Handling partial writes and fin-only writes in QUIC.
This handles an API problem for sending a fin with no associated data.

This can happen at the GFE if we're dechunking HTTP, have
sent all bytes read from a backend, and then we get a 0 byte terminal chunk.

In this case, we have no way of telling when we write data if the fin was
consumed, or if the write was blocked at an underlying layer and the end of
data would otherwise never be communicated to the peer.

Merge internal change: 40340063

Review URL: https://chromiumcodereview.appspot.com/11820003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@176287 0039d316-1c4b-4281-b951-d872f2087c98