Remove HttpStream::GetResponseInfo, which is not currently used.
HttpStream::GetResponseInfo returned a pointer to an object not owned
by the HttpStream, so could result in returning freed memory, in the
case the stream outlives its client.  This happens in the case an
HttpResponseBodyDrainer is used.

As the method isn't used anywhere, seems a good idea to get rid of it.

BUG=none

Review URL: https://codereview.chromium.org/303443009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@273629 0039d316-1c4b-4281-b951-d872f2087c98