QUIC -track how long it has taken to finish handshake once we start waiting for reading of QUIC server information from disk cache.
We could use this data to compare total time taken if we were to cancel
the disk cache read vs waiting for the read to complete.

Added a histogram to track the ratio of the time spent waiting to load
QUIC server information from disk cache to the min rtt. Logged when
session is closed.

R=asvitkine@chromium.org, rch@chromium.org
BUG=434936

Review URL: https://codereview.chromium.org/732353004

Cr-Commit-Position: refs/heads/master@{#305505}