relnote: not serializing invalid headers going from HTTP->SPDY. Behind FLAGS_spdy_strip_invalid_headers.
Prior to this CL we had a lovely pattern where you could take a set of
HTTP headers, serliaze to SPDY, and then serializing back to HTTP would
fail because we were writing out headers we would not read.

This should fix this once and for all.

Merge internal change: 96583692

R=rch@chromium.org

Review URL: https://codereview.chromium.org/1250373003

Cr-Commit-Position: refs/heads/master@{#340367}