Only MarkTrailersConsumed when actually sending trailers notification
MarkTrailersConsumed is currently done before the
notification is posted asynchronously. However, there might
be a pending read  before that notification is executed but
after  MarkTrailersConsumed, which leads to delegate
mistakenly to assume that the trailers are consumed.
This CL changes so that we MarkTrailersConsumed when we
actually will immediately notify delegate.

This CL also patches some internal trailers changes in
quic_spdy_stream.cc and quic_spdy_client_stream.cc.

BUG=586207

Review URL: https://codereview.chromium.org/1776423004

Cr-Commit-Position: refs/heads/master@{#380799}