Replace key_exchange_info with key_exchange_group.
Narrow to key_exchange_group in net::SSLInfo. The one place where
key_exchange_info remains is serialization code which sadly must filter
out garbage values.

This removes the DHE group size histogram which was the last consumer of
the non-ECDHE values. With DHE behind a fallback and now removed, there
is no need for this histogram.

BUG=639421

Review-Url: https://codereview.chromium.org/2349713004
Cr-Commit-Position: refs/heads/master@{#419753}