Avoid heap allocations in IPAddress
Use a helper class to represent the sequence of bytes in an IP address.
A vector<uint8_t> would be simpler but incurs heap allocation, so
IPAddressBytes uses a fixed size array.

IPAddress construction and destruction is showing up in cronet profiles
consuming roughly 5% of CPU. With this fix, it drops below .1%.

Also switch call sites which create IPAddress objects to avoid using std::vector and instead use base::StackVector.

Review-Url: https://codereview.chromium.org/2881673002
Cr-Commit-Position: refs/heads/master@{#474780}