Start refactor of QuicCryptoStream
QuicCryptoClientStream and QuicCryptoServerStream are refactored to delegate
their implementations to a CryptoClientStreamDelegate or
CryptoServerStreamDelegate, and the existing implementations are moved to new
subclasses of the new delegates.

The major work left to do in this refactor is for QuicCryptoStream to no longer
implement CryptoFramerVisitorInterface and to remove the OnHandshakeMessage
method from CryptoClientStreamDelegate and CryptoServerStreamDelegate
interfaces.

Merge internal change: 161549072

Change-Id: I3919d067028cdacd939c64be4362d4645d7e7e65
Reviewed-on: https://chromium-review.googlesource.com/565978
Commit-Queue: Nick Harper <nharper@chromium.org>
Reviewed-by: Ryan Hamilton <rch@chromium.org>
Cr-Commit-Position: refs/heads/master@{#486185}