Add request-level CT metrics
We already have per-connection CT metrics for percent of connections that are
CT-compliant (both overall and of connections which are required to be
CT-compliant). This CL adds per-request metrics, which may give a different
picture: for example, it's possible that overall compliance might look high as a
percentage of overall connections but low as a percentage of requests, because
one non-compliant connection might be responsible for a disproportionate number
of requests.

Bug: 772534
Cq-Include-Trybots: master.tryserver.chromium.android:android_cronet_tester;master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: I7c54dc7621588505ac5c30cf9d8d704cb6b59b59
Reviewed-on: https://chromium-review.googlesource.com/769295
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Reviewed-by: Matt Mueller <mattm@chromium.org>
Commit-Queue: Emily Stark <estark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#519015}