Remove support for the unused disable_hpack_dynamic_table connection option Merge internal change: 200745921
Clean up the QUIC end-to-end-test GetTestParams loop and comments.
Merge internal change: 200748036

Only run the stateless reject versions of QUIC end-to-end tests for a single test. Currently the end-to-end tests are parameterized in such a way that we run 60 different versions of each of the 80 tests in end_to_end_tests.cc. This is a *ton* of tests and contributes to net_unittests in Chromium taking a very long time.

One of the dimensions that these tests are parameterized by is stateless reject support:
* supported by the client?
* supported by the server?
* using cheap stateless rejects?
But this only affects the *handshake* and does not affect subsequent operations. So there's really no need to run all 80 tests with all the combinations of stateless reject-ness. This CL changes things so that only a single test runs with all the various stateless reject configurations and cuts the end_to_end_test runtime in half.
Merge internal change: 200763662

Change-Id: I83ac58d0b8d34692d9db309cb45cfd7c6c64622b
Bug: 806109
Reviewed-on: https://chromium-review.googlesource.com/1102895
Commit-Queue: Ryan Hamilton <rch@chromium.org>
Reviewed-by: Zhongyi Shi <zhongyi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#567786}