Remove net::ProxyDelegate as a per-request parameter of net::ProxyResolutionService.
Previously methods on ProxyResolutionService took an optional ProxyDelegate pointer. This was problematic as not all callers were providing the correct value, and plumbing it down was inconvenient.

Now the ProxyDelegate is instead an internal detail of the ProxyResolutionService.

Bug: 793071
Cq-Include-Trybots: luci.chromium.try:linux_mojo
Change-Id: I2eed185bc42cdc0bfcb945ad037fe2df52416198
Reviewed-on: https://chromium-review.googlesource.com/1208465
Reviewed-by: Matt Mueller <mattm@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Commit-Queue: Eric Roman <eroman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#589950}