Mock WriteHeaders on QuicSpdyStream instead of QuicSpdySession.
In current Google QUIC, HEADERS are sent on a dedicated headers frame
that is owned by QuicSpdySession.  So QuicSpdyStream::WriteHeaders()
calls QuicSpdySession::WriteHeaders(), which is virtual and is mocked in
tests.  However, in IETF QUIC, HEADERS will be sent on the
request/response stream, so writing headers is more naturally done
within QuicSpdyStream instead of QuicSpdySession.  This CL does not move
HEADERS to the request/response streams yet, but prepares tests for it
by:
* introducing virtual QuicSpdyStream::WriteHeadersImpl(),
* mocking it for tests instead of mocking
  QuicSpdySession::WriteHeaders(),
* renaming QuicSpdySession::WriteHeaders()
  to WriteHeadersOnHeadersStream(),
* introducing QuicSpdySession::WriteHeadersOnHeadersStreamImpl().

In QuicSimpleServerSessionTests a test implementation of QuicSpdySession
is used with regular QuicSpdyStreams.
MockQuicSpdySession::WriteHeaders() used to prevent actual writes, but
QuicSpdyStream::WriteHeadersImpl() is not overriden and writes do happen
on the headers stream.  gMock expectations are added accordingly.
Testing for the stream id is lost in the process, but stream id is
tested by QuicConnection::SendStreamData() expectations anyway.  In
other tests expectations are changed from the session object to the
stream object.

This CL lands server side changes 232878862 and 233946983 by bnc.

Patch Set 1: server change 232878862.

Patch Set 2: server change 233946983.

Patch Set 3: Chromium changes.

Change-Id: Id12c9691abc9a69d99c1dad854fcd929db8a6c3f
Reviewed-on: https://chromium-review.googlesource.com/c/1473132
Reviewed-by: Ryan Hamilton <rch@chromium.org>
Commit-Queue: Bence Béky <bnc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#632431}