Make the QuicParams more self-documenting
With the introduction of QuicParams, it's now possible to
default-construct a QuicParams. However, it's not
self-documenting what those defaults are. As many of them
are actually constants from the header themselves, align
the class to initialize in-place, making it more obvious.

This also updates the various time-delta based parameters
of QuicParams to use base::TimeDelta, making it easier to
ensure safe math, while freeing callers from needing to
know about whether the units were _seconds or
_milliseconds. Several members of the class already
followed this pattern, this just aligns all of them.

Bug: None
Change-Id: Id1661af6736740dd470d713843d37b2af9b7134a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1707095
Reviewed-by: Ryan Hamilton <rch@chromium.org>
Reviewed-by: Nick Harper <nharper@chromium.org>
Commit-Queue: Ryan Sleevi <rsleevi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#678442}