Replace GetList().{emplace,push}_back with Append in //net
This change replaces usages of base::Value::GetList() followed by either
emplace_back() or push_back() with base::Value::Append(). This is
because of the upcoming change to GetList() to return a base::span
instead, which does not support either emplace_back() or push_back().

This is a completely mechanical change. Steps to reproduce:
- sed -i 's/GetList().push_back/Append/g'
- sed -i 's/GetList().emplace_back/Append/g'
- git cl format

Bug: 646113
Change-Id: I36d53904d2be2d95d410ef88e3c973884a7777c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1796428
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Commit-Queue: Jan Wilken Dörrie <jdoerrie@chromium.org>
Cr-Commit-Position: refs/heads/master@{#695306}