Implement QuicTransportClient
QuicTransportClient is a top-level API for establishing and using QuicTransport
connection.  It handles DNS and proxy resolution, and creates all of the
necessary infrastructure (socket, certificate verifier, etc) for the underlying
QUIC session.

Currently, this is only covered by the basic end-to-end test.  An in-depth unit
test covering this would require QuicTestPacketMaker integration, which would
make this CL too large in scope.

Bug: 1011392
Change-Id: I19f37aed23440acb6a8eaa655c3af9170eade2a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1970141
Reviewed-by: Ryan Hamilton <rch@chromium.org>
Commit-Queue: Victor Vasiliev <vasilvv@chromium.org>
Cr-Commit-Position: refs/heads/master@{#726893}