Respect Alt-Svc order from server
We used to respect the server's Alt-Svc order for the IETF format
but not the legacy gQUIC format. This CL changes that so we respect
the server's ordering in all cases.

Note that this CL has the disadvantage of loading versions in an
incorrect sorted order the first time it will be run. However,
that will be fixed the moment we get a new Alt-Svc header from
the server, so it's an acceptable loss: the first connection
might used a valid non-preferred version once.

Bug: 1063060
Change-Id: I11851068fb13ab81ceeff6cebb0870e874d8bf3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2136407
Reviewed-by: Zhongyi Shi <zhongyi@chromium.org>
Commit-Queue: Zhongyi Shi <zhongyi@chromium.org>
Commit-Queue: David Schinazi <dschinazi@chromium.org>
Auto-Submit: David Schinazi <dschinazi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#757157}