Ignore legacy Alt-Svc format for versions that do not support it
This CL fixes an oversight from when
ParsedQuicVersion::SupportsGoogleAltSvcFormat was added.
We already never send the legacy format with new versions,
this CL makes it so we do not parse it with those versions.
We've checked with our partners that deploy gQUIC and they
have confirmed that this change will not cause issues for them.

Change-Id: I6957f91e8bbca9bd1134bbc2f4979c91e990435f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2204822
Commit-Queue: Nick Harper <nharper@chromium.org>
Reviewed-by: Nick Harper <nharper@chromium.org>
Auto-Submit: David Schinazi <dschinazi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#769804}